refactor: cortex cli as client - communicate with API server via cortexjs
