+++ cortex-js/src/infrastructure/commanders/usecases/cli.usecases.module.ts
@@ -10,7 +10,6 @@ import { AssistantsModule } from '@/usecases/assistants/assistants.module';
 import { MessagesModule } from '@/usecases/messages/messages.module';
 import { FileManagerModule } from '@/infrastructure/services/file-manager/file-manager.module';
 import { PSCliUsecases } from './ps.cli.usecases';
-import { BenchmarkCliUsecases } from './benchmark.cli.usecases';
 import { TelemetryModule } from '@/usecases/telemetry/telemetry.module';
 import { DownloadManagerModule } from '@/infrastructure/services/download-manager/download-manager.module';
 import { EnginesModule } from '@/usecases/engines/engines.module';
@@ -29,17 +28,7 @@ import { EnginesModule } from '@/usecases/engines/engines.module';
     DownloadManagerModule,
     EnginesModule,
   ],
-  providers: [
-    ModelsCliUsecases,
-    ChatCliUsecases,
-    PSCliUsecases,
-    BenchmarkCliUsecases,
-  ],
-  exports: [
-    ModelsCliUsecases,
-    ChatCliUsecases,
-    PSCliUsecases,
-    BenchmarkCliUsecases,
-  ],
+  providers: [ModelsCliUsecases, ChatCliUsecases, PSCliUsecases],
+  exports: [ModelsCliUsecases, ChatCliUsecases, PSCliUsecases],
 })
 export class CliUsecasesModule {}
